// src/modules/expense-invoice/enums/expense-invoice-status.enum.ts
export enum EXPENSE_INVOICE_STATUS {
  Nonexistent = 'expense.invoice.status.non_existent',
  Draft = 'expense.invoice.status.draft',
  Sent = 'expense.invoice.status.sent',
  Validated = 'expense.invoice.status.validated',
  Paid = 'expense.invoice.status.paid',
  PartiallyPaid = 'expense.invoice.status.partially_paid',
  Unpaid = 'expense.invoice.status.unpaid',
  Expired = 'expense.invoice.status.expired',
  Archived = 'quotation.status.archived',
}